+++ .gitignore
@@ -1,3 +1,4 @@
+/config.json
 /.idea
 /.vscode
 
@@ -58,3 +59,4 @@ typings/
 .yarn-integrity
 
 # Temp ignore 
+config.json
