+++ app.json
@@ -1,5 +1,5 @@
 {
-  "name": "RaffaMusic",
+  "name": "EvoBot",
   "description": "🤖 Discord Music Bot built with discord.js",
   "repository": "https://github.com/eritislami/evobot",
   "logo": "https://i.imgur.com/JFxgbWH.png",
