+++ commands/nowplaying.js
@@ -1,4 +1,4 @@
-const createBar = require("string-progressbar");
+const { splitBar } = require("string-progressbar");
 const { MessageEmbed } = require("discord.js");
 
 const { LOCALE } = require("../util/EvobotUtil");
@@ -28,7 +28,7 @@ module.exports = {
         "\u200b",
         new Date(seek * 1000).toISOString().substr(11, 8) +
           "[" +
-          createBar(song.duration == 0 ? seek : song.duration, seek, 20)[0] +
+          splitBar(song.duration == 0 ? seek : song.duration, seek, 20)[0] +
           "]" +
           (song.duration == 0 ? " ◉ LIVE" : new Date(song.duration * 1000).toISOString().substr(11, 8)),
         false
