+++ commands/pause.js
