--- commands/rename.jsx
@@ -1,9 +1,0 @@
-const Discord = require("discord.js");
-
-module.exports = {
-  name: "rename",
-  description: 'Rename nama orang (DIBUAT SAMA RIZKI)',
-  async execute(message, args) {
-    
-    }
-  }
