+++ index.js
@@ -4,8 +4,7 @@
 const { Client, Collection } = require("discord.js");
 const { readdirSync } = require("fs");
 const { join } = require("path");
-const { TOKEN, PREFIX, LOCALE } = require("./util/EvobotUtil");
-const path = require("path");
+const { TOKEN, PREFIX } = require("./util/EvobotUtil");
 const i18n = require("i18n");
 
 const client = new Client({
@@ -21,9 +20,10 @@ const cooldowns = new Collection();
 const escapeRegex = (str) => str.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
 
 i18n.configure({
-  locales: ["ar", "de", "en", "es", "fr", "it", "ko", "nl", "pl", "pt_br", "ru", "sv", "tr", "zh_cn", "zh_tw"],
-  directory: path.join(__dirname, "locales"),
+  locales: ["ar", "de", "en", "es", "fr", "it", "ko", "nl", "pl", "pt_br", "ru", "sv", "th", "tr", "vi", "zh_cn", "zh_sg", "zh_tw"],
+  directory: join(__dirname, "locales"),
   defaultLocale: "en",
+  retryInDefaultLocale: true,
   objectNotation: true,
   register: global,
 
@@ -50,7 +50,7 @@ i18n.configure({
  */
 client.on("ready", () => {
   console.log(`${client.user.username} ready!`);
-  client.user.setActivity(`Subscribe Raffa Dian | -help`, { type: "WATCHING" });
+  client.user.setActivity(`Bot Punya Raffa`, { type: "PLAYING" });
 });
 client.on("warn", (info) => console.log(info));
 client.on("error", console.error);
@@ -108,6 +108,6 @@ client.on("message", async (message) => {
     command.execute(message, args);
   } catch (error) {
     console.error(error);
-    message.reply(i18n.__("common.errorCommend")).catch(console.error);
+    message.reply(i18n.__("common.errorCommand")).catch(console.error);
   }
 });
