+++ locales/ar.json
@@ -161,6 +161,6 @@
     "disabled": "**مقفل**",
     "errorNotChannel": "تحتاج إلى الانضمام إلى قناة صوتية أولاً!",
     "cooldownMessage": "يرجى الانتظار {time} ثانية أخرى قبل إعادة استخدام الأمر `{name}`.",
-    "errorCommend": "حدث خطأ أثناء تنفيذ هذا الأمر."
+    "errorCommand": "حدث خطأ أثناء تنفيذ هذا الأمر."
   }
 }
