+++ locales/en.json
@@ -161,6 +161,6 @@
     "disabled": "**disabled**",
     "errorNotChannel": "You need to join a voice channel first!",
     "cooldownMessage": "please wait {time} more second(s) before reusing the `{name}` command.",
-    "errorCommend": "There was an error executing that command."
+    "errorCommand": "There was an error executing that command."
   }
 }
