+++ locales/ru.json
@@ -161,6 +161,6 @@
     "disabled": "**отключено**",
     "errorNotChannel": "Сначала вам нужно присоединиться к голосовому каналу!",
     "cooldownMessage": "Пожалуйста, подождите {time} секунды перед повторным использованием `{name}` команды.",
-    "errorCommend": "При выполнении этой команды произошла ошибка."
+    "errorCommand": "При выполнении этой команды произошла ошибка."
   }
 }
