+++ locales/sv.json
@@ -161,7 +161,7 @@
       "disabled": "**Inaktiverad**",
       "errorNotChannel": "Du måste gå med i en röstkanal först!",
       "cooldownMessage": "vänta {time} sekunder till innan du återanvänder kommandot `{name}`.",
-      "errorCommend": "Det uppstod ett fel vid körningen av kommandot."
+      "errorCommand": "Det uppstod ett fel vid körningen av kommandot."
     }
   }
