+++ locales/th.json
@@ -1,0 +1,167 @@
+{
+    "clip": {
+      "description": "เล่นคลิปเสียง",
+      "usagesReply": "การใช้งาน: /clip <ขื่อ>",
+      "errorQueue": "ไม่สามารถเล่นคลิปเพราะมีคิวที่ใช้งานอยู่.",
+      "errorNotChannel": "คุณต้องเข้าร่วมช่องเสียงก่อน!"
+    },
+    "clips": {
+      "description": "รายการคลิปทั้งหมด"
+    },
+    "help": {
+      "description": "แสดงคำสั่งและคำอธิบายทั้งหมด",
+      "embedTitle": "{botname} ช่วย",
+      "embedDescription": "รายการคำสั่งทั้งหมด"
+    },
+    "invite": {
+      "description": "ส่งลิงค์เชิญบอท"
+    },
+    "loop": {
+      "description": "เปิดเพลงลูป",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "result": "ลูปคือตอนนี้ {loop}"
+    },
+    "lyrics": {
+      "description": "รับเนื้อเพลงสำหรับเพลงที่กำลังเล่นอยู่ในปัจจุบัน",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "lyricsNotFound": "ไม่พบเนื้อเพลง {title}.",
+      "embedTitle": "{title} - เนื้อเพลง"
+    },
+    "move": {
+      "description": "ย้ายเพลงไปรอบๆ ในคิว",
+      "errorNotQueue": "ไม่มีคิว.",
+      "usagesReply": "การใช้งาน: {prefix}move <เลขคิว>",
+      "result": "<@{author}> 🚚 ย้ายเพลง **{title}** ไป {index} ในคิว."
+    },
+    "nowplaying": {
+      "description": "แสดงเล่นเพลงตอนนี้",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "embedTitle": "กำลังเล่น",
+      "live": " ◉ LIVE",
+      "timeRemaining": "เวลาที่เหลืออยู่: {time}"
+    },
+    "pause": {
+      "description": "หยุดเพลงที่กำลังเล่นอยู่ในปัจจุบัน",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "result": "<@{author}> ⏸ หยุดเพลง."
+    },
+    "ping": {
+      "description": "แสดงปิงเฉลี่ยของบอท",
+      "result": "📈 ปิงเฉลี่ยต่อ API: {ping} ms"
+    },
+    "play": {
+      "description": "เล่นเพลงจาก YouTube หรือ SoundCloud",
+      "errorNotChannel": "คุณต้องเข้าร่วมช่องเสียงก่อน!",
+      "errorNotInSameChannel": "คุณต้องอยู่ในช่องเดียวกันกับ {user}",
+      "usageReply": "การใช้งาน: {prefix}play <ชื่อวิดีโอ หรือ ลิงค์จากYouTube | ลิงค์ SoundCloud>",
+      "missingPermissionConnect": "ไม่สามารถเชื่อมต่อกับช่องสัญญาณเสียงการอนุญาตที่ขาดหายไป",
+      "missingPermissionSpeak": "ฉันไม่สามารถพูดในช่องสัญญาณเสียงนี้ให้แน่ใจว่าฉันมีสิทธิ์ที่เหมาะสม!",
+      "queueAdded": "✅ **{title}** ถูกเพิ่มไปยังคิวโดย <@{author}>",
+      "cantJoinChannel": "ไม่สามารถเข้าร่วมช่อง: {error}",
+      "queueEnded": "❌ คิวเพลงสิ้นสุดลง.",
+      "queueError": "ข้อผิดพลาด: {error}",
+      "startedPlaying": "🎶 เริ่มเล่นเพลง: **{title}** {url}",
+      "skipSong": "<@{author}> ⏩ ข้ามเพลง",
+      "pauseSong": "<@{author}> ⏸ หยุดเพลง.",
+      "resumeSong": "<@{author}> ▶ ดำเนินการต่อเพลงต่อ!",
+      "unmutedSong": "<@{author}> 🔊 เปิดเสียงเพลง!",
+      "mutedSong": "<@{author}> 🔇 ปิดเสียงเพลง!",
+      "decreasedVolume": "<@{author}> 🔉 ลดระดับเสียงเพลงปริมาณตอนนี้คือ {volume}%",
+      "increasedVolume": "<@{author}> 🔊 เพิ่มระดับเสียงเพลงปริมาณตอนนี้คือ {volume}%",
+      "loopSong": "<@{author}> ลูปคือตอนนี้ {loop}",
+      "stopSong": "<@{author}> ⏹ หยุดเพลง!",
+      "leaveChannel": "ออกจากช่องเสียง...",
+      "songNotFound": "ไม่พบเสียง"
+    },
+    "playlist": {
+      "description": "เล่นเพลย์ลิสต์จาก YouTube",
+      "usagesReply": "การใช้งาน: {prefix}playlist <ชื่อเพลย์ลิสต์ จากYouTube หรือ ลิงค์เพลย์ลิสต์>",
+      "errorNotChannel": "คุณต้องเข้าร่วมช่องเสียงก่อน!",
+      "errorNotInSameChannel": "คุณต้องอยู่ในช่องเดียวกันกับ {user}",
+      "missingPermissionConnect": "ไม่สามารถเชื่อมต่อกับช่องสัญญาณเสียงการอนุญาตที่ขาดหายไป",
+      "missingPermissionSpeak": "ฉันไม่สามารถพูดในช่องสัญญาณเสียงนี้ให้แน่ใจว่าฉันมีสิทธิ์ที่เหมาะสม!",
+      "errorNotFoundPlaylist": "ไม่พบเพลย์ลิสต์ :(",
+      "fetchingPlaylist": "⌛ ดึงเพลย์ลิสต์...",
+      "playlistCharLimit": "\nเพลย์ลิสต์ใหญ่กว่าขีด จำกัด ของอักขระ...",
+      "startedPlaylist": "<@{author}> เริ่มเพลย์ลิสต์",
+      "cantJoinChannel": "ไม่สามารถเข้าร่วมช่อง: {error}"
+    },
+    "pruning": {
+      "description": "สลับการตัดแต่งกิ่งของข้อความบอท",
+      "errorWritingFile": "มีข้อผิดพลาดในการเขียนไฟล์.",
+      "result": "การตัดแต่งกิ่งข้อความคือ {result}"
+    },
+    "queue": {
+      "description": "แสดงคิวเพลงและตอนนี้เล่น.",
+      "missingPermissionMessage": "การอนุญาตที่ขาดหายไปในการจัดการข้อความหรือเพิ่มปฏิกิริยา",
+      "errorNotQueue": "❌ **ไม่มีอะไรเล่นในเซิร์ฟเวอร์นี้**",
+      "currentPage": "หน้าปัจจุบัน - ",
+      "embedTitle": "คิวเพลง\n",
+      "embedCurrentSong": "**เพลงปัจจุบัน - [{title}]({url})**\n\n{info}"
+    },
+    "remove": {
+      "description": "ลบเพลงจากคิว",
+      "errorNotQueue": "ไม่มีคิว.",
+      "usageReply": "การใช้งาน: {prefix}remove <เลขคิว>",
+      "result": "<@{author}> ❌ ลบเพลง **{title}** จากคิว."
+    },
+    "resume": {
+      "description": "ดำเนินการต่อเล่นเพลงต่อในเพลงปัจจุบัน",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "resultNotPlaying": "<@{author}> ▶ ดำเนินการเล่นเพลงต่อ!",
+      "errorPlaying": "คิวไม่หยุดชั่วคราว."
+    },
+    "search": {
+      "description": "ค้นหาและเลือกวิดีโอเพื่อเล่น",
+      "usageReply": "การใช้งาน: {prefix}{name} <ชื่อวิดีโอ>",
+      "errorAlreadyCollector": "ตัวรวบรวมข้อความใช้งานอยู่ในช่องนี้แล้ว.",
+      "errorNotChannel": "คุณต้องเข้าร่วมช่องเสียงก่อน!",
+      "resultEmbedTtile": "**ตอบกลับด้วยหมายเลขเพลงที่คุณต้องการเล่น**",
+      "resultEmbedDesc": "ผลลัพธ์สำหรับ: {search}"
+    },
+    "shuffle": {
+      "description": "สุ่มคิว",
+      "errorNotQueue": "ไม่มีคิว.",
+      "result": "<@{author}> 🔀 สุ่มคิว"
+    },
+    "skip": {
+      "description": "ข้ามเพลงที่กำลังเล่นอยู่ในปัจจุบัน",
+      "errorNotQueue": "ไม่มีอะไรเล่น ที่ฉันสามารถข้ามให้คุณได้.",
+      "result": "<@{author}> ⏭ ข้ามเพลง"
+    },
+    "skipto": {
+      "description": "ข้ามไปที่หมายเลขคิวที่เลือก",
+      "usageReply": "การใช้งาน: {prefix}{name} <เลขคิว>",
+      "errorNotQueue": "ไม่มีคิว",
+      "errorNotValid": "คิวเป็นเพียง {length} เพลงยาว!",
+      "result": "<@{author}> ⏭ ข้ามไป {arg} songs"
+    },
+    "stop": {
+      "description": "หยุดเพลง",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "result": "<@{author}> ⏹ หยุดเพลง!"
+    },
+    "uptime": {
+      "description": "ตรวจสอบเวลาทำงาน",
+      "result": "เวลาต่อไป: `{days} วัน(s),{hours} ชั่วโมง, {minutes} นาที, {seconds} วินาที`"
+    },
+    "volume": {
+      "description": "เปลี่ยนปริมาณการเล่นเพลงในปัจจุบัน",
+      "errorNotQueue": "ไม่มีอะไรเล่น.",
+      "errorNotChannel": "คุณต้องเข้าร่วมช่องเสียงก่อน!",
+      "currentVolume": "🔊 ปริมาณปัจจุบันคือ: **{volume}%**",
+      "errorNotNumber": "โปรดใช้ตัวเลขเพื่อตั้งค่าระดับเสียง.",
+      "errorNotValid": "โปรดใช้ตัวเลขระหว่าง 0 - 100.",
+      "result": "ระดับเสียงที่ตั้งค่าเป็น: **{arg}%**"
+    },
+    "common": {
+      "on": "**เปิด**",
+      "off": "**ปิด**",
+      "enabled": "**เปิดใช้งาน**",
+      "disabled": "**ปิดใช้งาน**",
+      "errorNotChannel": "คุณต้องเข้าร่วมช่องเสียงก่อน!",
+      "cooldownMessage": "โปรดรอ {time} ไม่วินาที(s) ก่อนที่จะนำกลับมาใช้ใหม่ `{name}` คำสั้ง.",
+      "errorCommand": "มีข้อผิดพลาดในการดำเนินการ คำสั้ง."
+    }
+  }
+
