+++ package.json
@@ -12,17 +12,17 @@
     "node": ">=14.0.0"
   },
   "dependencies": {
-    "@discordjs/opus": "^0.5.0",
+    "@discordjs/opus": "^0.5.3",
     "array-move": "^3.0.1",
     "discord.js": "^12.5.1",
-    "ffmpeg-static": "^4.2.3",
-    "i18n": "^0.13.2",
+    "ffmpeg-static": "^4.4.0",
+    "i18n": "^0.13.3",
     "lyrics-finder": "^21.0.5",
     "simple-youtube-api": "^5.1.1",
     "soundcloud-downloader": "^0.2.3",
-    "string-progressbar": "^1.0.3",
-    "ytdl-core": "^4.4.5",
-    "ytdl-core-discord": "^1.2.5"
+    "string-progressbar": "^1.0.4",
+    "ytdl-core": "^4.9.1",
+    "ytdl-core-discord": "^1.3.1"
   },
   "scripts": {
     "start": "node index.js"
